Applied Access Control
Applied some access control the the Assessments and Submissions
controllers.
Simple check when a User goes to edit something. If they don't own it,
they can't get to the edit page for it.

Additionally, the Submissions Create and Index pages were slightly
updated.
Index page removed the create link. Have to access the page from a link
on an Assessment.
Create page, removed the select box for an Assessment.

+++ src/Assessment-Management-System/Controllers/AssessmentsController.cs
@@ -98,7 +98,15 @@ namespace Assessment_Management_System.Controllers
                 return NotFound();
             }
 
+            var user = await _manager.GetUserAsync(HttpContext.User);
+
             var assessment = await _context.Assessment.SingleOrDefaultAsync(m => m.ID == id);
+
+            if (user.Id != assessment.teacherID)
+            {
+                return RedirectToAction("Index");
+            }
+
             if (assessment == null)
             {
                 return NotFound();
@@ -117,6 +125,12 @@ namespace Assessment_Management_System.Controllers
             if (id != assessment.ID)
             {
                 return NotFound();
+            }
+
+            var user = await _manager.GetUserAsync(HttpContext.User);
+            if (user.Id != assessment.teacherID)
+            {
+                return RedirectToAction("Index");
             }
 
             if (ModelState.IsValid)
@@ -152,6 +166,13 @@ namespace Assessment_Management_System.Controllers
             }
 
             var assessment = await _context.Assessment.SingleOrDefaultAsync(m => m.ID == id);
+
+            var user = await _manager.GetUserAsync(HttpContext.User);
+            if (user.Id != assessment.teacherID)
+            {
+                return RedirectToAction("Index");
+            }
+
             if (assessment == null)
             {
                 return NotFound();
@@ -167,6 +188,12 @@ namespace Assessment_Management_System.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var assessment = await _context.Assessment.SingleOrDefaultAsync(m => m.ID == id);
+            var user = await _manager.GetUserAsync(HttpContext.User);
+            if (user.Id != assessment.teacherID)
+            {
+                return RedirectToAction("Index");
+            }
+
             _context.Assessment.Remove(assessment);
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
